Cl/refactor prep for facets (#42)
* refactor mappers
* FacetResult count is long not int

+++ Dfe.Data.SearchPrototype/SearchForEstablishments/Models/FacetResult.cs
@@ -13,14 +13,14 @@ public class FacetResult
     /// <summary>
     /// The number of records that belong to this facet value
     /// </summary>
-    public int? Count { get; }
+    public long? Count { get; }
 
     /// <summary>
     /// Constructor
     /// </summary>
     /// <param name="value">The value of the facet result</param>
     /// <param name="count">The number of records that belong to this facet value</param>
-    public FacetResult(string value, int? count)
+    public FacetResult(string value, long? count)
     {
         Value = value;
         Count = count;
